test_app.py
#!/usr/bin/env python3
"""
Simple test script for Vertex AI Image Recognition Tester
Tests basic functionality without requiring Google Cloud setup
"""

import os
import sys
import tempfile
from unittest.mock import patch, MagicMock

def test_imports():
    """Test that all required modules can be imported"""
    try:
        from app import app
        print("✅ All imports successful")
        return True
    except ImportError as e:
        print(f"❌ Import error: {e}")
        return False

def test_app_creation():
    """Test that the Flask app can be created"""
    try:
        from app import app
        assert app is not None
        assert hasattr(app, 'routes')
        print("✅ Flask app creation successful")
        return True
    except Exception as e:
        print(f"❌ App creation error: {e}")
        return False

def test_mock_predictions():
    """Test that mock predictions work correctly"""
    try:
        from app import get_mock_predictions
        predictions = get_mock_predictions()
        assert len(predictions) > 0
        assert all('displayName' in pred for pred in predictions)
        assert all('confidence' in pred for pred in predictions)
        assert all('bbox' in pred for pred in predictions)
        print("✅ Mock predictions working correctly")
        return True
    except Exception as e:
        print(f"❌ Mock predictions error: {e}")
        return False

def test_image_processing():
    """Test basic image processing functionality"""
    try:
        from PIL import Image
        import numpy as np
        
        # Create a test image
        test_image = Image.new('RGB', (100, 100), color='red')
        
        # Test basic operations
        assert test_image.size == (100, 100)
        assert test_image.mode == 'RGB'
        
        print("✅ Image processing libraries working")
        return True
    except Exception as e:
        print(f"❌ Image processing error: {e}")
        return False

def main():
    """Run all tests"""
    print("🧪 Running Vertex AI Image Recognition Tester tests...\n")
    
    tests = [
        test_imports,
        test_app_creation,
        test_mock_predictions,
        test_image_processing
    ]
    
    passed = 0
    total = len(tests)
    
    for test in tests:
        try:
            if test():
                passed += 1
        except Exception as e:
            print(f"❌ Test {test.__name__} failed with exception: {e}")
        print()
    
    print(f"📊 Test Results: {passed}/{total} tests passed")
    
    if passed == total:
        print("🎉 All tests passed! The application is ready to run.")
        print("\nTo start the application:")
        print("1. Set up your environment variables (see .env.example)")
        print("2. Run: python run.py")
        return True
    else:
        print("⚠️  Some tests failed. Please check the errors above.")
        return False

if __name__ == '__main__':
    success = main()
    sys.exit(0 if success else 1)